refactor: convert canned response callback
