+++ apps/meteor/ee/app/canned-responses/server/index.ts
@@ -5,7 +5,6 @@ await License.onLicense('canned-responses', async () => {
 	await import('./permissions');
 	await import('./hooks/onRemoveAgentDepartment');
 	await import('./hooks/onSaveAgentDepartment');
-	await import('./hooks/onMessageSentParsePlaceholder');
 	await import('./methods/saveCannedResponse');
 	await import('./methods/removeCannedResponse');
